% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_descriptive_statistics.R
\name{apa.desc}
\alias{apa.desc}
\title{Report descriptive statistics for a set of values}
\usage{
apa.desc(..., dv = NULL, data = NULL, show.mean = NULL,
  show.sd = NULL, show.se = NULL, show.conf.interval = NULL,
  show.N = NULL, number.decimals = NULL)
}
\arguments{
\item{...}{one or more dplyr commands (separated by commas) to select rows. Eg: alcohol=="2 Pints", gender=="Female"}

\item{dv}{Name of the dependent variable column}

\item{data}{Project data frame name}

\item{show.mean}{Show mean (Bool)}

\item{show.sd}{Show standard deviation ((Bool))}

\item{show.se}{Show standard error (Bool)}

\item{show.conf.interval}{Show confidence interval (Bool)}

\item{show.N}{Show number of cases (Bool)}
}
\value{
R Markdown text
}
\description{
Report descriptive statistics for a set of values
}
\examples{
#2-way ANOVA Example
library(apaTables) #load apaTables to access goggles

#Main Effect Means: Gender
apa.desc(goggles, gender=="Female", dv = attractiveness)
apa.desc(goggles, gender=="Male", dv = attractiveness)

#Main Effect Means: Alcohol
apa.desc(goggles, alcohol=="None", dv = attractiveness)
apa.desc(goggles, alcohol=="2 Pints", dv = attractiveness)
apa.desc(goggles, alcohol=="4 Pints", dv = attractiveness)

#Cell Mean: Female, 2 Pints
apa.desc(goggles, alcohol=="2 Pints", gender=="Female", dv = attractiveness)
}
