% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_test.R
\name{apa.t.test}
\alias{apa.t.test}
\title{Report descriptive statistics for a set of values}
\usage{
apa.t.test(t_test_object, show.mean.difference = TRUE,
  show.statistic = NULL, show.conf.interval = NULL,
  number.decimals = NULL, number.decimals.p = NULL)
}
\arguments{
\item{show.conf.interval}{Show confidence interval (Bool. Default TRUE)}

\item{...}{one or more dplyr commands (separated by commas) to select rows. Eg: alcohol=="2 Pints", gender=="Female"}

\item{dv}{Name of the dependent variable column}

\item{data}{Project data frame name}

\item{show.mean}{Show mean (Bool. Default TRUE)}

\item{show.sd}{Show standard deviation (Bool. Default TRUE)}

\item{show.se}{Show standard error (Bool. Default FALSE)}

\item{show.N}{Show number of cases (Bool. Default TRUE)}
}
\value{
R Markdown text
}
\description{
Report descriptive statistics for a set of values
}
\examples{
#2-way ANOVA Example
library(apaTables) #load apaTables to access goggles
}
