% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaText.R
\docType{package}
\name{apaText}
\alias{apaText}
\title{Create R Markdown Text for Results in the Style of the American Psychological Association (APA)}
\description{
Create APA style text from analyses for use within R Markdown documents. Descriptive statistics, confidence intervals, and cell sizes are reported.
\tabular{ll}{
Package: \tab apaText\cr
Type: \tab Package\cr
Version: \tab 0.1.7\cr
Date: \tab 2023-05-23\cr
License: \tab MIT\cr
}
}
\author{
\tabular{ll}{
Author: \tab David J. Stanley \email{dstanley@uoguelph.ca}\cr
Maintainer: \tab David J. Stanley \email{dstanley@uoguelph.ca}
}
}
