% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_utility.R
\name{set.apa.default.options}
\alias{set.apa.default.options}
\title{Create apaText default options for showing confidence intervals etc.. These options will be used unless overridden by local function arguments}
\usage{
set.apa.default.options()
}
\value{
A list with options object for apaText
}
\description{
Create apaText default options for showing confidence intervals etc.. These options will be used unless overridden by local function arguments
}
\examples{
# You must create an object called apa.default.options
# for options to be used, as per below.

apa.options <- set.apa.default.options()
}
