% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_t_test.R
\name{apa.t.test}
\alias{apa.t.test}
\title{Report descriptive statistics for a set of values}
\usage{
apa.t.test(t_test_object, show.mean.difference = TRUE,
  show.statistic = NULL, show.conf.interval = NULL,
  number.decimals = NULL, number.decimals.p = NULL)
}
\arguments{
\item{t_test_object}{Object returned from t.test function}

\item{show.mean.difference}{Show mean difference (Bool. Default TRUE)}

\item{show.statistic}{Show t-value (Bool. Default TRUE)}

\item{show.conf.interval}{Show CI for mean difference (Bool. Default TRUE)}

\item{number.decimals}{Number of decimals used in output (excluding p-value)}

\item{number.decimals.p}{Number of decimals used in p-value output}
}
\value{
R Markdown text
}
\description{
Report descriptive statistics for a set of values
}
\examples{
# Example here
}
